requirements.txt
dash
dash-bootstrap-components
folium
branca
pandas
scikit-learn
requests
python-dotenv
json